# perf: move path.slice inside iterateFirst